+++ src/android/com/synconset/cordovahttp/CordovaHttpGet.java
@@ -1,60 +1,48 @@
 /**
  * A HTTP plugin for Cordova / Phonegap
  */
-package com.synconset;
+package com.synconset.cordovahttp;
 
-import android.util.Log;
-
-import com.github.kevinsawicki.http.HttpRequest;
-import com.github.kevinsawicki.http.HttpRequest.HttpRequestException;
-
-import java.io.File;
 import java.net.UnknownHostException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.Map;
 
 import javax.net.ssl.SSLHandshakeException;
 
 import org.apache.cordova.CallbackContext;
-import org.apache.cordova.file.FileUtils;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
-public class CordovaHttpDownload extends CordovaHttp implements Runnable {
-    private String filePath;
-    
-    public CordovaHttpDownload(String urlString, Map<?, ?> params, Map<String, String> headers, CallbackContext callbackContext, String filePath) {
+import com.github.kevinsawicki.http.HttpRequest;
+import com.github.kevinsawicki.http.HttpRequest.HttpRequestException;
+
+class CordovaHttpGet extends CordovaHttp implements Runnable {
+    public CordovaHttpGet(String urlString, JSONObject params, JSONObject headers, CallbackContext callbackContext) {
         super(urlString, params, headers, callbackContext);
-        this.filePath = filePath;
     }
-    
+
     @Override
     public void run() {
         try {
-            HttpRequest request = HttpRequest.get(this.getUrlString(), this.getParams(), true);
+            HttpRequest request = HttpRequest.get(this.getUrlString(), this.getParamsMap(), false);
+
             this.setupSecurity(request);
             request.acceptCharset(CHARSET);
-            request.headers(this.getHeaders());
+            request.headers(this.getHeadersMap());
+
             int code = request.code();
-            
+            String body = request.body(CHARSET);
             JSONObject response = new JSONObject();
+
             this.addResponseHeaders(request, response);
             response.put("status", code);
+
             if (code >= 200 && code < 300) {
-                URI uri = new URI(filePath);
-                File file = new File(uri);
-                request.receive(file);
-                JSONObject fileEntry = FileUtils.getFilePlugin().getEntryForFile(file);
-                response.put("file", fileEntry);
+                response.put("data", body);
                 this.getCallbackContext().success(response);
             } else {
-                response.put("error", "There was an error downloading the file");
+                response.put("error", body);
                 this.getCallbackContext().error(response);
             }
-        } catch(URISyntaxException e) {
-            this.respondWithError("There was an error with the given filePath");
         } catch (JSONException e) {
             this.respondWithError("There was an error generating the response");
         } catch (HttpRequestException e) {