+++ src/android/com/synconset/cordovahttp/CordovaHttpHead.java
@@ -1,18 +1,9 @@
 /**
  * A HTTP plugin for Cordova / Phonegap
  */
-package com.synconset;
+package com.synconset.cordovahttp;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.net.UnknownHostException;
-import java.util.Map;
-
-import javax.net.ssl.HttpsURLConnection;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.HostnameVerifier;
 
 import javax.net.ssl.SSLHandshakeException;
 
@@ -20,27 +11,29 @@ import org.apache.cordova.CallbackContext;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import android.util.Log;
-
 import com.github.kevinsawicki.http.HttpRequest;
 import com.github.kevinsawicki.http.HttpRequest.HttpRequestException;
- 
-public class CordovaHttpHead extends CordovaHttp implements Runnable {
-    public CordovaHttpHead(String urlString, Map<?, ?> params, Map<String, String> headers, CallbackContext callbackContext) {
+
+class CordovaHttpHead extends CordovaHttp implements Runnable {
+    public CordovaHttpHead(String urlString, JSONObject params, JSONObject headers, CallbackContext callbackContext) {
         super(urlString, params, headers, callbackContext);
     }
-    
+
     @Override
     public void run() {
         try {
-            HttpRequest request = HttpRequest.head(this.getUrlString(), this.getParams(), true);
+            HttpRequest request = HttpRequest.head(this.getUrlString(), this.getParamsMap(), true);
+
             this.setupSecurity(request);
             request.acceptCharset(CHARSET);
-            request.headers(this.getHeaders());
+            request.headers(this.getHeadersMap());
+
             int code = request.code();
             JSONObject response = new JSONObject();
+
             this.addResponseHeaders(request, response);
             response.put("status", code);
+
             if (code >= 200 && code < 300) {
                 // no 'body' to return for HEAD request
                 this.getCallbackContext().success(response);