+++ src/android/com/synconset/cordovahttp/CordovaHttpPlugin.java
@@ -1,39 +1,26 @@
 /**
  * A HTTP plugin for Cordova / Phonegap
  */
-package com.synconset;
+package com.synconset.cordovahttp;
 
 import java.io.BufferedInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.HttpURLConnection;
-import java.security.GeneralSecurityException;
-import java.security.KeyStore;
-import java.security.cert.Certificate;
-import java.security.cert.CertificateException;
-import java.security.cert.CertificateFactory;
-import java.security.cert.X509Certificate;
-import java.util.Iterator;
-import java.util.ArrayList;
-import java.util.HashMap;
 
-import javax.net.ssl.HttpsURLConnection;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManagerFactory;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.HostnameVerifier;
+import java.security.GeneralSecurityException;
+
+import java.util.ArrayList;
 
 import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaInterface;
 import org.apache.cordova.CordovaPlugin;
 import org.apache.cordova.CordovaWebView;
+
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import android.content.res.AssetManager;
-import android.util.Base64;
-import android.util.Log;
 
 import com.github.kevinsawicki.http.HttpRequest;
 
@@ -51,25 +38,23 @@ public class CordovaHttpPlugin extends CordovaPlugin {
             String urlString = args.getString(0);
             JSONObject params = args.getJSONObject(1);
             JSONObject headers = args.getJSONObject(2);
-            HashMap<?, ?> paramsMap = this.getMapFromJSONObject(params);
-            HashMap<String, String> headersMap = this.getStringMapFromJSONObject(headers);
-            CordovaHttpGet get = new CordovaHttpGet(urlString, paramsMap, headersMap, callbackContext);
+            CordovaHttpGet get = new CordovaHttpGet(urlString, params, headers, callbackContext);
+
             cordova.getThreadPool().execute(get);
         } else if (action.equals("head")) {
             String urlString = args.getString(0);
             JSONObject params = args.getJSONObject(1);
             JSONObject headers = args.getJSONObject(2);
-            HashMap<?, ?> paramsMap = this.getMapFromJSONObject(params);
-            HashMap<String, String> headersMap = this.getStringMapFromJSONObject(headers);
-            CordovaHttpHead head = new CordovaHttpHead(urlString, paramsMap, headersMap, callbackContext);
+            CordovaHttpHead head = new CordovaHttpHead(urlString, params, headers, callbackContext);
+
             cordova.getThreadPool().execute(head);
         } else if (action.equals("post")) {
             String urlString = args.getString(0);
             JSONObject params = args.getJSONObject(1);
-            JSONObject headers = args.getJSONObject(2);
-            HashMap<?, ?> paramsMap = this.getMapFromJSONObject(params);
-            HashMap<String, String> headersMap = this.getStringMapFromJSONObject(headers);
-            CordovaHttpPost post = new CordovaHttpPost(urlString, paramsMap, headersMap, callbackContext);
+            String serializerName = args.getString(2);
+            JSONObject headers = args.getJSONObject(3);
+            CordovaHttpPost post = new CordovaHttpPost(urlString, params, serializerName, headers, callbackContext);
+
             cordova.getThreadPool().execute(post);
         } else if (action.equals("enableSSLPinning")) {
             try {
@@ -82,30 +67,30 @@ public class CordovaHttpPlugin extends CordovaPlugin {
             }
         } else if (action.equals("acceptAllCerts")) {
             boolean accept = args.getBoolean(0);
+
             CordovaHttp.acceptAllCerts(accept);
             callbackContext.success();
         } else if (action.equals("validateDomainName")) {
             boolean accept = args.getBoolean(0);
+
             CordovaHttp.validateDomainName(accept);
             callbackContext.success();
         } else if (action.equals("uploadFile")) {
             String urlString = args.getString(0);
             JSONObject params = args.getJSONObject(1);
             JSONObject headers = args.getJSONObject(2);
-            HashMap<?, ?> paramsMap = this.getMapFromJSONObject(params);
-            HashMap<String, String> headersMap = this.getStringMapFromJSONObject(headers);
             String filePath = args.getString(3);
             String name = args.getString(4);
-            CordovaHttpUpload upload = new CordovaHttpUpload(urlString, paramsMap, headersMap, callbackContext, filePath, name);
+            CordovaHttpUpload upload = new CordovaHttpUpload(urlString, params, headers, callbackContext, filePath, name);
+
             cordova.getThreadPool().execute(upload);
         } else if (action.equals("downloadFile")) {
             String urlString = args.getString(0);
             JSONObject params = args.getJSONObject(1);
             JSONObject headers = args.getJSONObject(2);
-            HashMap<?, ?> paramsMap = this.getMapFromJSONObject(params);
-            HashMap<String, String> headersMap = this.getStringMapFromJSONObject(headers);
             String filePath = args.getString(3);
-            CordovaHttpDownload download = new CordovaHttpDownload(urlString, paramsMap, headersMap, callbackContext, filePath);
+            CordovaHttpDownload download = new CordovaHttpDownload(urlString, params, headers, callbackContext, filePath);
+
             cordova.getThreadPool().execute(download);
         } else {
             return false;
@@ -148,27 +133,5 @@ public class CordovaHttpPlugin extends CordovaPlugin {
         } else {
             CordovaHttp.enableSSLPinning(false);
         }
-    }
-
-    private HashMap<String, String> getStringMapFromJSONObject(JSONObject object) throws JSONException {
-        HashMap<String, String> map = new HashMap<String, String>();
-        Iterator<?> i = object.keys();
-
-        while (i.hasNext()) {
-            String key = (String)i.next();
-            map.put(key, object.getString(key));
-        }
-        return map;
-    }
-
-    private HashMap<String, Object> getMapFromJSONObject(JSONObject object) throws JSONException {
-        HashMap<String, Object> map = new HashMap<String, Object>();
-        Iterator<?> i = object.keys();
-
-        while(i.hasNext()) {
-            String key = (String)i.next();
-            map.put(key, object.get(key));
-        }
-        return map;
     }
 }